requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.13.2
    # via sandbox (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via sandbox (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via argon2-cffi-bindings
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.1
    # via sandbox (pyproject.toml)
fastapi-cli==0.0.5
    # via fastapi
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   sandbox (pyproject.toml)
    #   fastapi
    #   sqlmodel
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyotp==2.9.0
    # via sandbox (pyproject.toml)
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.10
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.8.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.35
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.19
    # via sandbox (pyproject.toml)
starlette==0.37.2
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
uvicorn==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
